tests: fix for real CI on Linux by fix createFile() + fix catch(done)
The problem was not at all where I was looking for it. The error was
arising on CI because the autostart folder must not exists. Thus,
createFile() was wrongly throwing an error after calling mkdirp. Mkdirp
returns a string if a folder is created (the first one created if
recursive) or undefined if the folder already exists. In case of error,
it rejects the Promise.

Since Node fs can manage mkdir recursively itself, let's use the native
call and use the callback to manage the error/rejection if any.

Fix catch(done) to catch failures properly.

Signed-off-by: Alexandre Demers <alexandre.f.demers@gmail.com>